requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
av==11.0.0
braceexpand==0.1.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.4
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
ctranslate2==4.4.0
cycler==0.12.1
Cython==3.0.11
cytoolz==1.0.0
datasets==3.0.1
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
enum-compat==0.0.3
executing==2.1.0
faster-whisper==1.0.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.67.0
huggingface-hub==0.23.2
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
inflect==7.4.0
intervaltree==3.1.0
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
julius==0.2.7
jupyterlab_widgets==3.0.13
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.7
lazy_loader==0.4
Levenshtein==0.26.0
lhotse==1.27.0
librosa==0.10.2.post1
lightning==2.3.3
lightning-utilities==0.11.7
lilcom==1.8.0
lit==18.1.8
llvmlite==0.43.0
loguru==0.7.2
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nemo_toolkit==1.23.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.1.4
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.4.0.58
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
optuna==4.0.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
plac==1.4.3
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==17.0.0
pybind11==2.13.6
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytorch-metric-learning==2.6.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
resampy==0.4.3
rich==13.8.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
sox==1.5.0
soxr==0.5.0.post1
speechbrain==1.0.1
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.5.0
tokenizers==0.15.2
toolz==1.0.0
torch==2.0.0
torch-audiomentations==0.11.1
torch-model-archiver==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.0.0
torchmetrics==1.4.2
torchserve==0.12.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.39.3
triton==2.0.0
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==3.0.4
wget==3.2
whisperx @ git+https://github.com/m-bain/whisperx.git@9e3a9e0e38fcec1304e1784381059a0e2c670be5
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
youtokentome==1.0.6
